wikiwikialoha/requirements.txt
Flask
Flask-SocketIO
Flask-Cors
python-socketio[client]
eventlet
